hcm_project/requirements.txt
python==2.7
django==1.8
django-mptt
django-rest-framework
